chore: move cards to the folder
