+++ web/src/components/cards/providers/GitHubRepository.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {useTranslation} from 'react-i18next';
 import {FaGithub} from 'react-icons/fa';
 
-import {Card, Repeater, DedicatedAvailability} from './core';
+import {Card, Repeater, DedicatedAvailability} from '../core';
 
 const GithubCard: React.FC<{query: string}> = ({query}) => {
   const {t} = useTranslation();
