+++ web/src/components/cards/providers/Slack.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {useTranslation} from 'react-i18next';
 import {FaSlack} from 'react-icons/fa';
 
-import {Card, DedicatedAvailability, Repeater} from './core';
+import {Card, DedicatedAvailability, Repeater} from '../core';
 
 const SlackCard: React.FC<{query: string}> = ({query}) => {
   const {t} = useTranslation();
