+++ web/src/components/cards/providers/Spectrum.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {useTranslation} from 'react-i18next';
-import {Card, Repeater, DedicatedAvailability} from './core';
-import {SpectrumIcon} from '../Icons';
+import {Card, Repeater, DedicatedAvailability} from '../core';
+import {SpectrumIcon} from '../../Icons';
 
 const SpectrumCard: React.FC<{query: string}> = ({query}) => {
   const {t} = useTranslation();
