+++ web/src/components/cards/providers/Twitter.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import {useTranslation} from 'react-i18next';
 import {FaTwitter} from 'react-icons/fa';
 
-import {capitalize} from '../../util/text';
-import {Card, Repeater, DedicatedAvailability} from './core';
+import {capitalize} from '../../../util/text';
+import {Card, Repeater, DedicatedAvailability} from '../core';
 
 const TwitterCard: React.FC<{query: string}> = ({query}) => {
   const {t} = useTranslation();
